feat: not get exist img

+++ script/utils.py
@@ -292,7 +292,11 @@ def timestamp2date(ts: int):
     return dt_object.strftime("%Y-%m-%d")
 
 
-def wget_file(url, output):
+def wget_file(url: str, output):
+    if (
+        url.endswith(".png") or url.endswith(".jpg") or url.endswith(".jpeg")
+    ) and os.path.exists(output):
+        return
     os.system(f"wget {url} -O {output}")
